requirements.txt
PyQt6==6.6.0
influxdb==5.3.1
python-dotenv==1.0.0